+++ src/core/strings/string_stream.h
@@ -89,7 +89,7 @@ namespace string_stream
 
 	inline StringStream& operator<<(StringStream& s, const char* string)
 	{
-		array::push(s, string, strlen(string));
+		array::push(s, string, (uint32_t)strlen(string));
 		return s;
 	}
