+++ src/lua/lua_stack.h
@@ -176,7 +176,7 @@ struct LuaStack
 
 	int get_int(int i)
 	{
-		return CHECKINTEGER(L, i);
+		return (int)CHECKINTEGER(L, i);
 	}
 
 	StringId32 get_string_id(int i)
